requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
cffi==1.16.0
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.110.0
fonttools==4.50.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
install==1.3.5
joblib==1.3.2
jwcrypto==1.5.6
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
regex==2023.12.25
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
